incentive_server/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.0.6
django-admin-bootstrapped==2.5.6
django-bootstrap3==12.1.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-nested-inline==0.4.2
djangorestframework==3.11.0
Faker==4.1.0
Flask==1.1.2
future==0.15.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.4
pandas==1.0.4
psycopg2==2.8.5
Pygments==2.6.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2019.3
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
text-unidecode==1.3
threadpoolctl==2.1.0
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.0.1
django-crontab==0.7.1
celery==4.4.1
redis
